+++ Changes.md
@@ -20,11 +20,15 @@ refer to [https://github.com/msimerson/Mail-Toaster-6/commits/master](https://gi
 - vpopmail: remove qmailadmin SPAM_COMMAND
 - vpopmail: build updates
 - vpopmail: replace lighttpd with nginx
+- webmail: when nginx is the proxy...
+  - use ACME module for TLS
+  - proxy websockets to /watch
 - unifi: configure external mongodb when UNIFI_MONGODB_DSN
 - ci: added minimal permissions
 - ci: enabled FreeBSD basic tests
 - ci: set of kcov testing and codecov reporting
 - consistency: always quote variables, explicit exit codes
+- contrib/exorcise_maildrop: add -n dry-run mode
 - test: expanded ./test/mail-toaster.bats
 - test: added ./test/provision/*.bats